构建：锁定 C++20 编译器与标准库
在配置阶段检查 GCC、Clang、AppleClang 与 MSVC 最低版本，并探测 format、source_location 和 stop_token 标准库能力。

为一方翻译单元启用跨平台严格警告和可选 MSVC /analyze，通过正向路径聚合执行 clang-tidy，隔离 MOC、RCC、qwindowkit 与 vendored spdlog。

修正门禁发现的窄化转换、无效移动、额外复制和跨平台未使用状态，并补充 Qt 所有权与异常边界的局部静态分析说明。

+++ ZzFluentUI/tests/ZzFluentScreenshotTest.cpp
@@ -280,7 +280,7 @@ QImage zzBuildTextMask(
             const QRect textRect = zzAlignedTextRect(
                 label,
                 label->contentsRect(),
-                label->alignment(),
+                static_cast<int>(label->alignment()),
                 label->text());
             zzPaintMaskRect(&painter, zzMapToSurface(label, textRect, surface));
             ++coverage->labels;
@@ -793,6 +793,8 @@ public:
     }
 
 private Q_SLOTS:
+    // QApplication 接管 setStyle() 传入对象；静态分析器不了解 Qt 所有权。
+    // NOLINTBEGIN(clang-analyzer-cplusplus.NewDeleteLeaks)
     void initTestCase()
     {
         QLocale::setDefault(QLocale::c());
@@ -821,6 +823,7 @@ private Q_SLOTS:
         QApplication::setStyle(
             new ZzFluentUI::ZzFluentStyle(controller_.get(), fusion));
     }
+    // NOLINTEND(clang-analyzer-cplusplus.NewDeleteLeaks)
 
     void rendersThemes_data()
     {
